### 2025-06-17 #8418 @editorialbot
[PRE REVIEW]: Synthesizer: Synthetic Observables For Modern Astronomy
**Submitting author:** @WillJRoper (<a href="http://orcid.org/0000-0002-3257-8806">Will Roper</a>)
**Repository:** https://github.com/synthesizer-project/synthesizer
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.9.4
