+++ docs/joss_paper/synthesizer.bib
@@ -358,7 +358,9 @@ archivePrefix = {arXiv},
     Publisher = {IEEE},
     Title = {OpenMP: an industry standard API for shared-memory programming},
     Volume = {5},
-    Year = {1998}}
+    Year = {1998},
+    doi = {10.1109/99.660313},
+}
 
 @article{dense_basis,
 doi = {10.3847/1538-4357/ab2052},
