### 2025-06-17 #8418 @warrickball
Hi @WillJRoper, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. With our current backlog, this will likely take at least three months. Sorry, and thanks for your patience!

In the meantime, note that your paper is a bit long. Our docs on [what the paper should contain](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) say

> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

The main purpose of the paper is to motivate why the software package exists, which usually also requires a small description of the domain (as you have). We don't expect it to contain implementation details, though the references to the packages and datasets on which Synthesizer depends is good.

I don't expect you to get all the way down to 1000 words but it would be great if you could cut a few hundred ahead of the review. If you wish to check the word count again, comment on this issue with `@editorialbot check repository`. You can re-render the paper with `@editorialbot generate pdf`.

Conversely, if you can make some space for it, I think it would help to demonstrate what Synthesizer can do by including some.
