### 2025-06-18 #8418 @WillJRoper
Hi @warrickball, that's a shame, but thanks for the heads up! I've reached out to @JBorrow separately, but I'll ask the question here too: Does being waitlisted mean the scope review will be pushed back, or will that happen at some point in the meantime?

Apologies on the word count, I got carried away with addressing co-author comments and adding the related packages section. I then failed to double-check that the count hadn't exploded... I'll trim it down shortly. While I tried to avoid implementation details, I wanted to include the list of abstractions to link to the documentation (mainly so that someone reading the paper element would be able to quickly get to points of interest). Is this acceptable?
