### 2025-06-18 3ba3bd1
Cutting word count
