+++ docs/joss_paper/synthesizer.bib
@@ -589,4 +589,16 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2011ApJS..192....9T},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
+@article{unyt,
+  doi = {10.21105/joss.00809},
+  url = {https://doi.org/10.21105/joss.00809},
+  year  = {2018},
+  month = {aug},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {28},
+  pages = {809},
+  author = {Nathan J. Goldbaum and John A. ZuHone and Matthew J. Turk and Kacper Kowalik and Anna L. Rosen},
+  title = {unyt: Handle,  manipulate,  and convert data with units in Python},
+  journal = {Journal of Open Source Software}
+}
