### 2025-06-18 24ec176
Cutting word count
