+++ docs/joss_paper/paper.md
@@ -108,7 +108,7 @@ Synthesizer addresses these shortcomings by offering:
 
 Synthesizer's design facilitates apples-to-apples comparisons between simulations and observations [e.g. @FLARESXVIII], permits exhaustive tests of the impact of parameter choices [e.g. @LTU-ILI], enables the forward modelling of large datasets previously considered impractical [e.g. @LTU-LOVELL], and promotes open and reproducible science.
 
-# Implementation overview
+# Package overview
 
 Synthesizer is structured around a set of core abstractions. Here we give a brief outline of these abstractions and a link to the documentation for each.
 
@@ -124,43 +124,17 @@ Synthesizer is hosted on [GitHub](https://github.com/synthesizer-project/synthes
 
 ## Related packages
 
-There are various other related packages which either perform similar tasks as Synthesizer or can be used in harmony with Synthesizer to create end-to-end workflows. We highlight a handful of these packages below.
+Several packages either overlap with Synthesizer’s functionality or complement it in end-to-end workflows:
 
-- **Stellar population synthesis & Photoionisation**
-
-  - **BC03** [@bc03]: Classic stellar population synthesis code providing spectral energy distributions (SEDs) for single stellar populations with various initial mass functions (IMFs) and metallicities.
-  - **FSPS** [@fsps1; @fsps2]: Stellar population synthesis code offering flexible isochrone-based spectral generation for a variety of IMFs and metallicities.
-  - **BPASS** [@bpass]: Models binary stellar populations with detailed spectral outputs, crucial for UV and nebular emission studies.
-  - **Maraston** [@maraston05]: Provides high-resolution SEDs with alternative stellar evolution prescriptions, particularly useful for post-starburst galaxies.
-  - **Cloudy** [@cloudy]: Photoionisation and spectral synthesis code for computing emission-line and continuum processes in gas under various physical conditions.
-  - **MAPPINGS** [@MAPPINGS]: An alternative photoionisation code for computing emission-line spectra from ionised gas, particularly useful for AGN and star-forming regions.
-
-- **Monte Carlo radiative transfer**
-
-  - **SKIRT** [@SKIRT]: 3D dust radiative-transfer engine supporting arbitrary geometries, multi-wavelength photon packets, and variance-reduction techniques for high-fidelity galaxy and torus models.
-  - **Powderday** [@POWDERDAY]: Integrates FSPS, Hyperion [@hyperion], and yt [@YT] to perform Monte Carlo RT directly on hydrodynamic simulation outputs, automating grid preparation and execution.
-
-- **Point spread function & instrument models**
-
-  - **STPSF** [@stpsf]: Physical-optics simulator for JWST, Roman, and HST, modelling pupil masks, wavefront errors, and instrument-specific aberrations to generate realistic PSFs.
-  - **GalSim** [@galsim]: Versatile image-simulation toolkit offering analytic and empirical PSF models, shear and magnification operators, detector effects (e.g., charge diffusion), and realistic noise injection.
-
-- **Pre- and post-processing utilities**
-
-  - **Astropy** [@astropy]: Foundational library for astronomy, providing FITS I/O, WCS transformations, units system, coordinate conversions, and utility functions for photometry and statistics.
-  - **YT** [@YT]: Analysis and visualisation toolkit for astrophysical simulations, supporting adaptive mesh refinement (AMR) data, particle data, and structured grids, with a focus on large-scale cosmological simulations.
[…]
