### 2025-06-18 #8418 @WillJRoper
Word count has now been cut down. Note that the reported 1433 is the full markdown file. Without the header and Acknowledgements (most of which are a necessity) is now 956 words. I hope this is alright now!
