### 2025-06-18 930027a
Workflow with the tex artifact
