+++ src/synthesizer/components/__init__.py
@@ -1,0 +1,1 @@
+from synthesizer.components.component import Component
