### 2025-06-23 df1a7d9
Fixing accepts decortator
