### 2025-06-23 e58cdc2
Adding generator base class
