### 2025-06-23 2546878
Better base class name for dust emission
