### 2025-06-23 9ac2fa4
Moving scaling by emissions bolometric luminosity to scale_by
