+++ src/synthesizer/emission_models/base_model.py
@@ -49,6 +49,7 @@ import numpy as np
 from unyt import arcsecond, kpc, unyt_quantity
 
 from synthesizer import exceptions
+from synthesizer.emission_models.generators.generator import Generator
 from synthesizer.emission_models.operations import (
     Combination,
     Extraction,
@@ -128,14 +129,8 @@ class EmissionModel(Extraction, Generation, Transformation, Combination):
         dust_curve (emission_models.attenuation.*):
             The dust curve to apply.
         generator (EmissionModel):
-            The emission generation model. This must define a get_spectra
-            method.
-        lum_intrinsic_model (EmissionModel):
-            The intrinsic model to use deriving the dust luminosity when
-            computing dust emission.
-        lum_attenuated_model (EmissionModel):
-            The attenuated model to use deriving the dust luminosity when
-            computing dust emission.
+            The emission generation model. This must define a generate_lnu and
+            a generate_lines method.
         mask_attr (str):
             The component attribute to mask on.
         mask_thresh (unyt_quantity):
@@ -145,7 +140,12 @@ class EmissionModel(Extraction, Generation, Transformation, Combination):
         lam_mask (ndarray):
             The mask to apply to the wavelength array.
         scale_by (list):
-            A list of attributes to scale the spectra by.
+            A list of attributes or emission models to scale the output
+            emission by. If an element is a string, it should be the name of
+            an emitter attribute, this will be extracted from the emitter and
+            the output emission will be multiplied by this value. If an element
+            is an EmissionModel, the bolometric luminosity of that model's
+            output spectra will be used to multiply the output emission.
         post_processing (list):
             A list of post processing functions to apply to the emission after
             it has been generated. Each function must take a dict containing
@@ -180,8 +180,6 @@ class EmissionModel(Extraction, Generation, Transformation, Combination):
         igm=None,
         generator=None,
         transformer=None,
-        lum_intrinsic_model=None,
-        lum_attenuated_model=None,
         mask_attr=None,
         mask_thresh=None,
         mask_op=None,
@@ -235,12 +233,6 @@ class EmissionModel(Extraction, Generation, Transformation, Combination):
                 The transform to apply. This is also an alternative (but
                 less obvious) argument for passing a dust curve or IGM model
                 (both are transformers).
-            lum_intrinsic_model (EmissionModel):
-                The intrinsic model to use deriving the dust luminosity when
-                computing dust emission.
-            lum_attenuated_model (EmissionModel):
-                The attenuated model to use deriving the dust luminosity when
-                computing dust emission.
             mask_attr (str):
                 The component attribute to mask on.
             mask_thresh (unyt_quantity):
@@ -337,8 +329,6 @@ class EmissionModel(Extraction, Generation, Transformation, Combination):
             transformer=transformer,
             apply_to=apply_to,
             generator=generator,
-            lum_intrinsic_model=lum_intrinsic_model,
-            lum_attenuated_model=lum_attenuated_model,
         )
 
         # Initilaise the corresponding operation (also checks we have a
@@ -357,8 +347,6 @@ class EmissionModel(Extraction, Generation, Transformation, Combination):
                 else igm
             ),
             generator=generator,
-            lum_intrinsic_model=lum_intrinsic_model,
-            lum_attenuated_model=lum_attenuated_model,
             vel_shift=vel_shift,
         )
 
@@ -414,8 +402,6 @@ class EmissionModel(Extraction, Generation, Transformation, Combination):
         apply_to,
         transformer,
         generator,
-        lum_intrinsic_model,
-        lum_attenuated_model,
         vel_shift,
     ):
         """Initialise the correct parent operation.
@@ -435,12 +421,6 @@ class EmissionModel(Extraction, Generation, Transformation, Combination):
             generator (EmissionModel):
[…]
