### 2025-07-21 77023a1
fix missing mask attrs
