### 2025-08-05 8649e51
Added thread local images to marginly improve scaling
