+++ src/synthesizer/imaging/extensions/image.cpp
@@ -56,21 +56,6 @@ void populate_smoothed_image_serial(const double *pix_values,
                                     const double threshold, const int kdim,
                                     double *img, const int nimgs) {
 
-  /* Find the maximum kernel_cdim we'll need. We need this to preallocate the
-   * kernel we will populate for each particle. */
-  int max_kernel_cdim = 0;
-  for (int ind = 0; ind < npart; ind++) {
-    const double smooth_length = smoothing_lengths[ind];
-    const int delta_pix = ceil(smooth_length / res) + 1;
-    const int kernel_cdim_temp = 2 * delta_pix + 1;
-    if (kernel_cdim_temp > max_kernel_cdim) {
-      max_kernel_cdim = kernel_cdim_temp;
-    }
-  }
-
-  /* Allocate the particle kernel. */
-  double *part_kernel = new double[max_kernel_cdim * max_kernel_cdim];
-
   /* Loop over positions including the sed */
   for (int ind = 0; ind < npart; ind++) {
 
@@ -83,17 +68,25 @@ void populate_smoothed_image_serial(const double *pix_values,
     const int i = x / res;
     const int j = y / res;
 
+    /* If the smoothing length is less than the resolution just add the
+     * pixel value to the image. */
+    if (smooth_length <= res) {
+      for (int nimg = 0; nimg < nimgs; nimg++) {
+        /* Get the pixel index in the image (nimg, i, j). */
+        const int pix_ind = nimg * npix_x * npix_y + npix_y * i + j;
+
+        /* Add the pixel value to this image, pixel values are (Nimg, Npart)
+         * in shape. */
+        img[pix_ind] += pix_values[nimg * npart + ind];
+      }
+      continue;
+    }
+
     /* How many pixels are in the smoothing length? Add some buffer. */
     const int delta_pix = ceil(smooth_length / res) + 1;
 
     /* How many pixels along kernel axis? */
     const int kernel_cdim = 2 * delta_pix + 1;
-
-    /* Zero the part of the kernel we will use. */
-    memset(part_kernel, 0, kernel_cdim * kernel_cdim * sizeof(double));
-
-    /* Track the kernel sum for normalisation. */
-    double kernel_sum = 0;
 
     /* Compute the pixel indices for the kernel. */
     int ii_min = (i - delta_pix) < 0 ? 0 : (i - delta_pix);
@@ -110,10 +103,18 @@ void populate_smoothed_image_serial(const double *pix_values,
       /* Compute the x separation */
       const double x_dist = res * (ii + 0.5) - x;
 
+      /* Early exit if the x distance is too large. */
+      if (x_dist * x_dist > thresh2)
+        continue;
+
       for (int jj = jj_min; jj <= jj_max; jj++) {
 
         /* Compute the y separation */
         const double y_dist = res * (jj + 0.5) - y;
+
+        /* Early exit if the y distance is too large. */
+        if (y_dist * y_dist > thresh2)
+          continue;
 
         /* Compute the distance between the centre of this pixel
          * and the particle. */
@@ -123,45 +124,15 @@ void populate_smoothed_image_serial(const double *pix_values,
         if (rsqu > thresh2)
           continue;
 
-        /* Get the pixel coordinates in the kernel */
-        const int iii = ii - (i - delta_pix);
-        const int jjj = jj - (j - delta_pix);
-
         /* Calculate the impact parameter. */
         const double q = sqrt(rsqu) / smooth_length;
 
         /* Get the value of the kernel at q. */
         const int index = kdim * q;
-        const double kvalue = kernel[index];
+        const double kvalue =
+            kernel[index] / (smooth_length * smooth_length) * res * res;
 
[…]
