### 2025-08-05 4c0c470
Removing redundant code
