### 2025-08-05 83fbb66
Actually use the thread img...
