+++ src/synthesizer/emission_models/tree_plot.py
@@ -1,0 +1,940 @@
+"""A submodule for plotting the tree structure of emission models."""
+
+import matplotlib.lines as mlines
+import matplotlib.patches as mpatches
+import matplotlib.pyplot as plt
+import numpy as np
+
+
+def get_tree_levels(models_dict, root):
+    """Get the levels of the tree.
+
+    Args:
+        models_dict (dict): Dictionary mapping model labels to model objects
+        root (str): The root of the tree to get the levels for
+
+    Returns:
+        tuple: (levels, links, extract_labels, masked_labels)
+            - levels (dict): Dictionary mapping levels to lists of model labels
+            - links (dict): Dictionary mapping source labels to lists of
+                (target, style) tuples
+            - extract_labels (set): Set of labels for models that are
+                extracting
+            - masked_labels (list): List of labels for models that are masked
+    """
+
+    def _assign_levels(
+        levels,
+        links,
+        extract_labels,
+        masked_labels,
+        model,
+        level,
+    ):
+        """Recursively assign levels to the models."""
+        # Get the model label
+        label = model.label
+
+        # Assign the level
+        levels[model.label] = max(levels.get(model.label, level), level)
+
+        # Define the links
+        if model._is_transforming:
+            links.setdefault(label, []).append((model.apply_to.label, "--"))
+        if model._is_combining:
+            links.setdefault(label, []).extend(
+                [(child.label, "-") for child in model._combine]
+            )
+        if model._is_dust_emitting or model._is_generating:
+            links.setdefault(label, []).extend(
+                [
+                    (
+                        model._lum_intrinsic_model.label
+                        if model._lum_intrinsic_model is not None
+                        else None,
+                        "dotted",
+                    ),
+                    (
+                        model._lum_attenuated_model.label
+                        if model._lum_attenuated_model is not None
+                        else None,
+                        "dotted",
+                    ),
+                ]
+            )
+
+        if model._is_masked:
+            masked_labels.append(label)
+        if model._is_extracting:
+            extract_labels.add(label)
+
+        # Recurse
+        for child in model._children:
+            (
+                levels,
+                links,
+                extract_labels,
+                masked_labels,
+            ) = _assign_levels(
+                levels,
+                links,
+                extract_labels,
+                masked_labels,
+                child,
+                level + 1,
+            )
+
+        return levels, links, extract_labels, masked_labels
+
+    # Get the root model
+    root_model = models_dict[root]
+
+    # Recursively assign levels
+    (
+        model_levels,
+        links,
+        extract_labels,
+        masked_labels,
+    ) = _assign_levels({}, {}, set(), [], root_model, 0)
+
+    # Unpack the levels
+    levels = {}
+    for label, level in model_levels.items():
+        levels.setdefault(level, []).append(label)
+
+    return levels, links, extract_labels, masked_labels
+
+
+def _get_adaptive_spacing(levels, links):
+    """Calculate adaptive spacing based on tree structure.
+
+    Args:
+        levels (dict): Dictionary mapping levels to lists of model labels
+        links (dict): Dictionary of links between models
+
+    Returns:
+        tuple: (xchunk, ychunk) horizontal and vertical spacing values
+    """
+    max_level_width = max(len(models) for models in levels.values())
+    max_label_length = max(
+        len(label)
+        for level_models in levels.values()
+        for label in level_models
+    )
+
+    # Base spacing on content and tree width
[…]
