+++ src/synthesizer/emission_models/base_model.py
@@ -42,9 +42,6 @@ Example usage::
 import copy
 import sys
 
-import matplotlib.lines as mlines
-import matplotlib.patches as mpatches
-import matplotlib.pyplot as plt
 import numpy as np
 from unyt import arcsecond, kpc, unyt_quantity
 
@@ -55,6 +52,7 @@ from synthesizer.emission_models.operations import (
     Generation,
     Transformation,
 )
+from synthesizer.emission_models.tree_plot import _plot_emission_tree
 from synthesizer.extensions.timers import tic, toc
 from synthesizer.imaging import ImageCollection
 from synthesizer.imaging.image_generators import (
@@ -1562,283 +1560,14 @@ class EmissionModel(Extraction, Generation, Transformation, Combination):
                 data=[child.label.encode("utf-8") for child in self._children],
             )
 
-    def _get_tree_levels(self, root):
-        """Get the levels of the tree.
-
-        Args:
-            root (str):
-                The root of the tree to get the levels for.
-
-        Returns:
-            levels (dict):
-                The levels of the models in the tree.
-            links (dict):
-                The links between models.
-            extract_labels (set):
-                The labels of models that are extracting.
-            masked_labels (list):
-                The labels of models that are masked.
-        """
-
-        def _assign_levels(
-            levels,
-            links,
-            extract_labels,
-            masked_labels,
-            model,
-            level,
-        ):
-            """Recursively assign levels to the models.
-
-            Args:
-                levels (dict):
-                    The levels of the models.
-                links (dict):
-                    The links between models.
-                extract_labels (set):
-                    The labels of models that are extracting.
-                masked_labels (list):
-                    The labels of models that are masked.
-                components (dict):
-                    The component each model acts on.
-                model (EmissionModel):
-                    The model to assign the level to.
-                level (int):
-                    The level to assign to the model.
-
-            Returns:
-                levels (dict):
-                    The levels of the models.
-                links (dict):
-                    The links between models.
-                extract_labels (set):
-                    The labels of models that are extracting.
-                masked_labels (list):
-                    The labels of models that are masked.
-                components (dict):
-                    The component each model acts on.
-            """
-            # Get the model label
-            label = model.label
-
-            # Assign the level
-            levels[model.label] = max(levels.get(model.label, level), level)
-
-            # Define the links
-            if model._is_transforming:
-                links.setdefault(label, []).append(
-                    (model.apply_to.label, "--")
-                )
-            if model._is_combining:
-                links.setdefault(label, []).extend(
-                    [(child.label, "-") for child in model._combine]
-                )
-            if model._is_dust_emitting or model._is_generating:
-                links.setdefault(label, []).extend(
-                    [
-                        (
-                            model._lum_intrinsic_model.label
-                            if model._lum_intrinsic_model is not None
-                            else None,
-                            "dotted",
-                        ),
-                        (
-                            model._lum_attenuated_model.label
-                            if model._lum_attenuated_model is not None
-                            else None,
-                            "dotted",
-                        ),
-                    ]
-                )
-
-            if model._is_masked:
-                masked_labels.append(label)
-            if model._is_extracting:
-                extract_labels.add(label)
-
-            # Recurse
-            for child in model._children:
-                (
-                    levels,
-                    links,
-                    extract_labels,
-                    masked_labels,
-                ) = _assign_levels(
-                    levels,
-                    links,
-                    extract_labels,
-                    masked_labels,
-                    child,
-                    level + 1,
-                )
-
[…]
