### 2025-08-15 b57b9a9
Implement kwarg resolution function
