### 2025-08-21 ab317f3
Fixing reduction overflow
