+++ src/synthesizer/extensions/reductions.cpp
@@ -6,7 +6,7 @@
 #include "timers.h"
 
 /**
- * @brief Reduce Npart spectra to integrated spectra.
+ * @brief Reduce Npart spectra to size_tegrated spectra.
  *
  * This is a serial version of the function.
  *
@@ -16,21 +16,21 @@
  * @param npart: The number of particles.
  */
 static void reduce_spectra_serial(double *spectra, double *part_spectra,
-                                  int nlam, int npart) {
+                                  int nlam, size_t npart) {
   /* Loop over particles. */
-  for (int p = 0; p < npart; p++) {
+  for (size_t p = 0; p < npart; p++) {
     /* Loop over wavelengths. */
     for (int ilam = 0; ilam < nlam; ilam++) {
+      size_t part_spec_ind = p * nlam + ilam;
       /* Use fused multiply-add to accumulate with better precision.
        * Equivalent to: += spec_val * weight, but with a single rounding. */
-      spectra[ilam] =
-          std::fma(part_spectra[p * nlam + ilam], 1.0, spectra[ilam]);
+      spectra[ilam] = std::fma(part_spectra[part_spec_ind], 1.0, spectra[ilam]);
     }
   }
 }
 
 /**
- * @brief Reduce Npart spectra to integrated spectra.
+ * @brief Reduce Npart spectra to size_tegrated spectra.
  *
  * This is a parallel version of the function.
  *
@@ -42,13 +42,14 @@ static void reduce_spectra_serial(double *spectra, double *part_spectra,
  */
 #ifdef WITH_OPENMP
 static void reduce_spectra_parallel(double *spectra, double *part_spectra,
-                                    int nlam, int npart, int nthreads) {
+                                    int nlam, size_t npart, int nthreads) {
   /* Loop over particles in parallel. */
 #if defined(_OPENMP) && _OPENMP >= 201511
 #pragma omp parallel for num_threads(nthreads) reduction(+ : spectra[ : nlam])
-  for (int p = 0; p < npart; p++) {
+  for (size_t p = 0; p < npart; p++) {
     for (int ilam = 0; ilam < nlam; ilam++) {
-      spectra[ilam] += part_spectra[p * nlam + ilam];
+      size_t part_spec_ind = p * nlam + ilam;
+      spectra[ilam] += part_spectra[part_spec_ind];
     }
   }
 #else // OpenMP < 4.5 or no array reduction support
@@ -57,9 +58,10 @@ static void reduce_spectra_parallel(double *spectra, double *part_spectra,
     // Thread-local accumulation to avoid false sharing and atomics
     std::vector<double> local(nlam, 0.0);
 #pragma omp for nowait schedule(static)
-    for (int p = 0; p < npart; p++) {
+    for (size_t p = 0; p < npart; p++) {
       for (int ilam = 0; ilam < nlam; ilam++) {
-        local[ilam] += part_spectra[p * nlam + ilam];
+        size_t part_spec_ind = p * nlam + ilam;
+        local[ilam] += part_spectra[part_spec_ind];
       }
     }
     // Merge
@@ -75,7 +77,7 @@ static void reduce_spectra_parallel(double *spectra, double *part_spectra,
 #endif
 
 /**
- * @brief Reduce Npart spectra to integrated spectra.
+ * @brief Reduce Npart spectra to size_tegrated spectra.
  *
  * This is a wrapper function that calls the serial or parallel version of the
  * function depending on the number of threads requested or whether OpenMP is
@@ -87,8 +89,8 @@ static void reduce_spectra_parallel(double *spectra, double *part_spectra,
  * @param npart: The number of particles.
  * @param nthreads: The number of threads to use.
  */
[…]
