### 2025-08-21 ad9c03b
npart should always be size_t
