+++ src/synthesizer/extensions/reductions.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // Reduction prototypes
-void reduce_spectra(double *spectra, double *part_spectra, int nlam, int npart,
+void reduce_spectra(double *spectra, double *part_spectra, int nlam, size_t npart,
                     int nthreads);

[…]
