### 2025-08-21 e9b4229
More int -> size_t for particle count fixes
