### 2025-09-30 893de14
Found rare 0 width warnign failure
