### 2025-09-30 b93751a
Adding compilation with debugging flags and disabling optimizations
