### 2025-09-30 1a58c61
Missing outside of cell catch
