### 2025-09-30 46e2058
Better handling of nested failures
