+++ src/synthesizer/imaging/extensions/image.cpp
@@ -221,12 +221,20 @@ build_balanced_work_list(struct cell *root, int nthreads,
  * @param out The output array to populate with the pixel values.
  * @param nimgs The number of images to populate.
  * @param pix_values The pixel values to use for each image.
+ * @param res The pixel resolution.
+ * @param npix_x The number of pixels along the x axis.
+ * @param npix_y The number of pixels along the y axis.
+ *
+ * @return ret Return code (0 = success, 1 = failure).
  */
-static void populate_pixel_recursive(const struct cell *c, double threshold,
-                                     int kdim, const double *kernel, int npart,
-                                     double *out, int nimgs,
-                                     const double *pix_values, const double res,
-                                     const int npix_x, const int npix_y) {
+static int populate_pixel_recursive(const struct cell *c, double threshold,
+                                    int kdim, const double *kernel, int npart,
+                                    double *out, int nimgs,
+                                    const double *pix_values, const double res,
+                                    const int npix_x, const int npix_y) {
+
+  /* Default to success. */
+  int ret = 0;
 
   /* Is the cell split? */
   if (c->split) {
@@ -241,8 +249,13 @@ static void populate_pixel_recursive(const struct cell *c, double threshold,
       }
 
       /* Recurse... */
-      populate_pixel_recursive(cp, threshold, kdim, kernel, npart, out, nimgs,
-                               pix_values, res, npix_x, npix_y);
+      int child_ret =
+          populate_pixel_recursive(cp, threshold, kdim, kernel, npart, out,
+                                   nimgs, pix_values, res, npix_x, npix_y);
+      if (child_ret != 0) {
+        /* Bubble up failure immediately. */
+        return 1;
+      }
     }
 
   } else {
@@ -273,9 +286,9 @@ static void populate_pixel_recursive(const struct cell *c, double threshold,
     int j_min = std::max(0, (int)floor(y_min_world / res));
     int j_max = std::min(npix_y, (int)ceil(y_max_world / res));
 
-    /* If we are outside the cell then just return. */
+    /* If we are outside the cell then just return success (nothing to do). */
     if (i_min >= npix_x || i_max <= 0 || j_min >= npix_y || j_max <= 0) {
-      return;
+      return 0;
     }
 
     /* Calculate the dimensions of our local image buffer. */
@@ -284,15 +297,15 @@ static void populate_pixel_recursive(const struct cell *c, double threshold,
 
     /* Safety check: ensure we have valid dimensions. */
     if (local_width <= 0 || local_height <= 0) {
-      PyErr_SetString(PyExc_ValueError,
-                      "Invalid local image dimensions calculated.");
-      return;
+      /* Do not touch Python here; simply signal failure. */
+      return 1;
     }
 
     /* Allocate and zero-initialize the local image buffer. This will store
      * contributions from all particles in this cell before we copy them
      * back to the global image. */
-    std::vector<double> local_img(local_width * local_height * nimgs, 0.0);
+    std::vector<double> local_img((size_t)local_width * local_height * nimgs,
+                                  0.0);
 
     /* Unpack the particles from this leaf cell. */
     struct particle *parts = c->particles;
@@ -326,7 +339,7 @@ static void populate_pixel_recursive(const struct cell *c, double threshold,
             for (int nimg = 0; nimg < nimgs; nimg++) {
               int local_idx =
                   local_i * local_height * nimgs + local_j * nimgs + nimg;
-              if (local_idx >= 0 && local_idx < local_img.size()) {
+              if (local_idx >= 0 && local_idx < (int)local_img.size()) {
                 local_img[local_idx] += pix_values[part->index * nimgs + nimg];
               }
             }
[…]
