### 2025-10-13 de4bd22
Initial pass at masking combinations
