### 2025-10-15 9017d4e
Added methods and dicts for line images
