+++ src/synthesizer/base_galaxy.py
@@ -79,6 +79,15 @@ class BaseGalaxy:
         self.images_noise_lnu = {}
         self.images_noise_fnu = {}
 
+        # Define the dictionaries to hold the line images (same structure as
+        # images)
+        self.line_images_lnu = {}
+        self.line_images_fnu = {}
+        self.line_images_psf_lnu = {}
+        self.line_images_psf_fnu = {}
+        self.line_images_noise_lnu = {}
+        self.line_images_noise_fnu = {}
+
         # Initialise the dictionary to hold instrument specific spectroscopy
         self.spectroscopy = {}
 
@@ -1595,6 +1604,406 @@ class BaseGalaxy:
 
         # Return the image at the root of the emission model
         return images[emission_model.label]
+
+    def get_line_images_luminosity(
+        self,
+        line_ids,
+        emission_model,
+        img_type="smoothed",
+        instrument=None,
+        kernel=None,
+        kernel_threshold=1,
+        nthreads=1,
+        limit_to=None,
+        resolution=None,
+        fov=None,
+        cosmo=None,
+    ):
+        """Make line ImageCollections from luminosities.
+
+        For Parametric Galaxy objects, images can only be smoothed. An
+        exception will be raised if a histogram is requested.
+
+        For Particle Galaxy objects, images can either be a simple
+        histogram ("hist") or an image with particles smoothed over
+        their SPH kernel.
+
+        Which images are produced is defined by the emission model. If any
+        of the necessary line data is missing for generating a particular
+        image, an exception will be raised.
+
+        The limit_to argument can be used if only a specific image is desired.
+
+        Note that black holes will never be smoothed and only produce a
+        histogram due to the point source nature of black holes.
+
+        All images that are created will be stored on the emitter (Stars,
+        BlackHole/s, or galaxy) under the line_images_lnu attribute. The line
+        image collection at the root of the emission model will also be
+        returned.
+
+        Args:
+            line_ids (list):
+                A list of line ids to include in the line images.
+            emission_model (EmissionModel):
+                The emission model to use to generate the line images.
+            img_type (str):
+                The type of image to be made, either "hist" -> a histogram, or
+                "smoothed" -> particles smoothed over a kernel for a particle
+                galaxy. Otherwise, only smoothed is applicable.
+            instrument (Instrument):
+                The instrument to use for the image. This can be None but if
+                not it will be used to label the images by instrument.
+            kernel (np.ndarray of float):
+                The values from one of the kernels from the kernel_functions
+                module. Only used for smoothed images.
+            kernel_threshold (float):
+                The kernel's impact parameter threshold (by default 1).
+            nthreads (int):
+                The number of threads to use in the tree search. Default is 1.
+            limit_to (str/list):
+                Optionally pass a single model label to limit image generation
+                to only that model.
+            resolution (unyt_quantity of float):
+                The size of a pixel.
+                (Ignoring any supersampling defined by psf_resample_factor)
+            fov (unyt_quantity of float):
+                The width of the image in image coordinates.
+            cosmo (astropy.cosmology):
+                The cosmology to use for the calculation of the luminosity
+                distance. Only needed for internal conversions from cartesian
+                to angular coordinates when an angular resolution is used.
+
+        Returns:
+            dict: A dictionary containing line ImageCollections keyed by
+                line_id.
+        """
+        # Ensure we aren't trying to make a histogram for a parametric galaxy
+        if self.galaxy_type == "Parametric" and img_type == "hist":
+            raise exceptions.InconsistentArguments(
+                "Parametric Galaxies can only produce smoothed images."
+            )
+
+        # Ensure we aren't trying to make an image for a particle galaxy
+        # without a per particle model
+        if self.galaxy_type == "Particle" and not emission_model.per_particle:
+            raise exceptions.InconsistentArguments(
[…]
