### 2025-10-15 8721617
Implemented line image generators
