### 2025-10-15 a5ff6b9
Missing dict methods on line implemented
