### 2025-10-15 0281a62
Line image factory method on emission model implemented
