+++ src/synthesizer/emission_models/base_model.py
@@ -59,6 +59,7 @@ from synthesizer.extensions.timers import tic, toc
 from synthesizer.imaging import ImageCollection
 from synthesizer.imaging.image_generators import (
     _generate_image_collection_generic,
+    _generate_line_image_collection_generic,
 )
 from synthesizer.photometry import PhotometryCollection
 from synthesizer.synth_warnings import deprecation, warn
@@ -3197,6 +3198,219 @@ class EmissionModel(Extraction, Generation, Transformation, Combination):
         toc("Generating all images", start)
 
         return images
+
+    def _get_line_images(
+        self,
+        line_ids,
+        instrument,
+        fov,
+        emitters,
+        img_type="smoothed",
+        line_images=None,
+        _is_related=False,
+        limit_to=None,
+        do_flux=False,
+        kernel=None,
+        kernel_threshold=1.0,
+        cosmo=None,
+        nthreads=1,
+        **kwargs,
+    ):
+        """Generate line images as described by the emission model.
+
+        This will create line images for all models in the emission model which
+        have saved lines for the requested line_ids, unless limit_to is set
+        to a specific model, in which case only that model will have line
+        images generated.
+
+        Args:
+            line_ids (list):
+                A list of line ids to include in the line images.
+            instrument (Instrument):
+                The instrument to use for the line image generation.
+            fov (float):
+                The field of view of the image in angular units.
+            emitters (dict):
+                The emitters to generate the line images for in the form of a
+                dictionary, {"stellar": <emitter>, "blackhole": <emitter>}.
+            img_type (str):
+                The type of image to generate. Options are "smoothed" or
+                "hist".
+            line_images (dict):
+                A dictionary of line images to add to. Used for recursive
+                calls.
+            _is_related (bool):
+                Are we generating related model line images?
+            limit_to (str, list):
+                If not None, defines specific model(s) to limit generation to.
+            do_flux (bool):
+                If True, generate from fluxes, if False from luminosities.
+            kernel (str):
+                The convolution kernel to use for image generation.
+            kernel_threshold (float):
+                The threshold for the convolution kernel.
+            cosmo (Cosmology):
+                The cosmology to use for image generation.
+            nthreads (int):
+                The number of threads to use for image generation.
+            **kwargs (dict):
+                Any additional keyword arguments.
+
+        Returns:
+            dict: A dictionary of line ImageCollections keyed by model label,
+                  then by line_id.
+        """
+        start = tic()
+
+        # We don't want to modify the original emission model
+        emission_model = copy.copy(self)
+
+        # If we haven't got a line_images dictionary yet we'll make one
+        if line_images is None:
+            line_images = {}
+
+        # Convert `limit_to` to a list if it is a string
+        if limit_to is not None:
+            limit_to = (
+                [limit_to] if isinstance(limit_to, str) else limit_to.copy()
+            )
+
+        # If we are limiting to a specific model/s and these are a combination
+        # model, we need to make sure we include the models they are
+        # combining.
+        _orig_limit_to = limit_to
+        if limit_to is not None:
+            _orig_limit_to = limit_to.copy()
+            for label in limit_to:
+                # Get this model
+                this_model = emission_model._models[label]
+
+                # If this is a combination model, add the models it is
+                # combining to the list
+                if this_model._is_combining:
+                    limit_to.extend([m.label for m in this_model.combine])
+
+            # Remove duplicates
+            limit_to = list(set(limit_to))
+
+        # Set up the list to collect all the line data
+        lines_data = {e: {} for e in emitters.keys()}
+
+        # Loop through all models and collect their line data
+        for label in emission_model._models.keys():
+            # Get this model
+            this_model = emission_model._models[label]
+
+            # If we are limiting to a specific model, skip all others
+            if limit_to is not None and label not in limit_to:
+                continue
+
[…]
