### 2025-10-29 9b0cc2a
added dust attenuation
