+++ src/synthesizer/emission_models/agn/unified_agn.py
@@ -23,6 +23,10 @@ Example usage::
 from unyt import deg
 
 from synthesizer.emission_models.base_model import BlackHoleEmissionModel
+from synthesizer.emission_models.models import (
+    AttenuatedEmission,
+    DustEmission,
+)
 from synthesizer.emission_models.transformers import (
     CoveringFraction,
     EscapingFraction,
@@ -67,21 +71,31 @@ class UnifiedAGN(BlackHoleEmissionModel):
         covering_fraction_blr="covering_fraction_blr",
         covered_fraction="covered_fraction",
         disc_transmission="combined",
-        label="intrinsic",
+        diffuse_dust_curve=None,
+        diffuse_dust_emission_model=None,
         **kwargs,
     ):
         """Initialize the UnifiedAGN model.
 
         Args:
-            nlr_grid (synthesizer.grid.Grid): The grid for the NLR.
-            blr_grid (synthesizer.grid.Grid): The grid for the BLR.
-            torus_emission_model (synthesizer.dust.EmissionModel): The dust
-                emission model to use for the torus.
-            covering_fraction_nlr (float): The covering fraction of the NLR.
-            covering_fraction_blr (float): The covering fraction of the BLR.
-            covered_fraction (float): The covering fraction of the disc.
-            disc_transmission (str): The disc transmission model.
-            label (str): The label for the model.
+            nlr_grid (synthesizer.grid.Grid):
+                The grid for the NLR.
+            blr_grid (synthesizer.grid.Grid):
+                The grid for the BLR.
+            torus_emission_model (synthesizer.dust.EmissionModel):
+                The dust emission model to use for the torus.
+            covering_fraction_nlr (float):
+                The covering fraction of the NLR.
+            covering_fraction_blr (float):
+                The covering fraction of the BLR.
+            covered_fraction (float):
+                The covering fraction of the disc.
+            disc_transmission (str):
+                The disc transmission model.
+            diffuse_dust_curve (synthesizer.emission_models.attenuation):
+                The dust attenuation curve for diffuse dust.
+            diffuse_dust_emission_model:
+                The diffuse dust emission model.
             **kwargs: Any additional keyword arguments to pass to the
                 BlackHoleEmissionModel.
         """
@@ -142,27 +156,81 @@ class UnifiedAGN(BlackHoleEmissionModel):
         # Get the torus emission model
         self.torus = self._make_torus(torus_emission_model, **kwargs)
 
-        # Create the final model
-        BlackHoleEmissionModel.__init__(
-            self,
-            label=label,
-            combine=(
-                self.disc,
-                self.nlr,
-                self.blr,
-                self.torus,
-            ),
-            related_models=(
-                self.disc_incident_isotropic,
-                self.disc_incident,
-                self.disc_averaged,
-                self.disc_averaged_without_torus,
-                self.disc,
-                self.nlr_continuum,
-                self.blr_continuum,
-            ),
-            **kwargs,
-        )
+        # If diffuse_dust_curve and diffuse_dust_emission_model provided then
+        # include these
+        if diffuse_dust_curve and diffuse_dust_emission_model:
+            # Calculate the intrinsic emission
+            self.intrinsic = BlackHoleEmissionModel(
+                label="intrinsic",
+                combine=(
+                    self.disc,
+                    self.nlr,
+                    self.blr,
+                    self.torus,
+                ),
+                **kwargs,
+            )
+
+            # Include attenuation from diffuse dust
+            # self.attenuated = self._add_diffuse_dust_attenuation(
+            # diffuse_dust_curve, **kwargs)
+            self.attenuated = AttenuatedEmission(
+                dust_curve=diffuse_dust_curve,
+                apply_to=self.intrinsic,
[…]
