### 2025-11-04 08ca72f
Adding base distribution abstract base class
