### 2025-11-04 2c1a6d7
Fixing return typing
