### 2025-11-04 d87b4be
Simplified distribution
