+++ src/synthesizer/distributions/distribution.py
@@ -7,6 +7,7 @@ used for sampling values in various contexts within the synthesizer framework.
 from abc import ABC, abstractmethod
 from typing import Tuple, Union
 
+import matplotlib.pyplot as plt
 import numpy as np
 from unyt import Unit, dimensionless, unyt_array
 
@@ -20,51 +21,6 @@ class Distribution(ABC):
     sampled from. Subclasses must implement the `sample` method to provide
     an interface for generating random samples according to the specific
     distribution.
-
-    Methods:
-        sample(size: int) -> list:
-            Generate random samples from the distribution.
-
-    Example:
-        >>> class NormalDistribution(Distribution):
-        ...     def __init__(self, mean: float, stddev: float):
-        ...         super().__init__(
-        ...             units=Unit("unitless"),
-        ...             required_parameters=("mean", "stddev"),
-        ...         )
-        ...         self.mean = mean
-        ...         self.stddev = stddev
-        ...     def distribution(self, x):
-        ...         # Define the normal distribution function here
-        ...         pass
-
-        >>> normal_dist = NormalDistribution(mean=0, stddev=1)
-        >>> samples = normal_dist.sample(size=1000)
-
-        >>> # Sample with parameters extracted from an object
-        >>> normal_dist = NormalDistribution(mean=None, stddev=None)
-        >>> stars = Stars(..., mean=10, stddev=3)
-        >>> extracted_params = normal_dist.extract_parameters(stars)
-        >>> samples = normal_dist.sample(size=1000, **extracted_params)
-
-        >>> # Mix fixed and extracted parameters
-        >>> normal_dist_fixed = NormalDistribution(mean=0, stddev=None)
-        >>> stars = Stars(..., stddev=2)
-        >>> extracted_params = normal_dist_fixed.extract_parameters(stars)
-        >>> samples = normal_dist_fixed.sample(size=1000, **extracted_params)
-
-        >>> # Extracting parameters from multiple objects
-        >>> normal_dist_multi = NormalDistribution(mean=None, stddev=None)
-        >>> stars = Stars(..., mean=5)
-        >>> m = EmissionModel(..., stddev=1)
-        >>> extracted_params = normal_dist_multi.extract_parameters(stars, m)
-
-        >>> # Extracting using list of objects with priority
-        >>> normal_dist = NormalDistribution(mean=None, stddev=None)
-        >>> s = Stars(..., mean=5)
-        >>> g = Gas(..., mean=10, stddev=2)
-        >>> extracted_params = normal_dist_priority.extract_parameters(s, g)
-        >>> samples = normal_dist.sample(size=1000, **extracted_params)
     """
 
     def __init__(
@@ -88,25 +44,21 @@ class Distribution(ABC):
         self.required_parameters = required_parameters
 
     @abstractmethod
-    def distribution(self, **kwargs) -> float:
+    def _distribution(self) -> float:
         """Define the probability distribution function.
 
         This method should be implemented by subclasses to define the
         specific probability distribution function.
 
-        Args:
-            **kwargs:
-                Keyword arguments required by the distribution.
-
         Returns:
-            The (unitless) value of the probability distribution function.
+            The (unitless) value of the probability distribution function
+            generated using the Distribution's attributes.
         """
         pass
 
     def sample(
         self,
-        size: int,
-        **kwargs,
+        nsamples: int,
     ) -> Union[unyt_array, np.ndarray]:
         """Generate random samples from the distribution.
 
@@ -114,7 +66,7 @@ class Distribution(ABC):
         probability distribution function.
 
         Args:
-            size (int):
+            nsamples (int):
                 The number of samples to generate.
             **kwargs:
                 Keyword arguments required by the distribution.
[…]
