### 2025-11-14 #8418 @grburgess
@TobiBu this seems really close to my own published JOSS work. Not sure why it’s needed or how I would contribute. Perhaps suggest to the authors to contribute to existing software?
