### 2025-11-14 #8418 @WillJRoper
Hi @grburgess, sorry you feel we have trodden on your toes here. I would like to point out that while there is crossover in overall goal (synthetic observations), I can see from a cursory glance that the codes could not be more fundamentally different in implementation, ethos, and scope. We set out with Synthesizer because we were unaware of any code that forward-modelled both parametric, SAM, and hydrodynamic simulation inputs and fulfilled our needs. Again, I am sorry you feel there is cross over here, but I assure you that was not our intention, and the two codes fundamentally do different things and can coexist.
