### 2025-11-17 #9436 @editorialbot
[REVIEW]: Synthesizer: Synthetic Observables For Modern Astronomy
**Submitting author:** @WillJRoper (<a href="http://orcid.org/0000-0002-3257-8806">Will Roper</a>)
**Repository:** https://github.com/synthesizer-project/synthesizer
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.0
