### 2025-11-18 #9436 @ConnorStoneAstro
Hi @WillJRoper , I'm starting my review and I see that Synthesizer is a pretty incredible package, very well done! The scope of the project is really broad, covering many different model and observation modes, I'm not aware of anything that covers this much range.

I am starting with the paper, it seems very well written and clearly defines the need that you are filling. However, looking at the "state of the field" and "references" elements of the review checklist, I think the current "Related packages" section could use a bit of work. It's a tough section to write since Synthesizer covers so much ground. Currently, it seems the section is focused on packages that could act as plugins to supplement synthesizer, though it does include other codes. I think the section would benefit from some clarity on which codes have plugin functionality online, which ones can interface but would require the user to do some work, and which are fully independent. There are also a number of missing citations from this section for codes that perform conceptually similar synthetic observation of simulated objects:

- GalaxyGenius: https://arxiv.org/pdf/2506.15060v1
- FORECAST: https://arxiv.org/pdf/2305.19166
- py-ananke: https://arxiv.org/pdf/2312.02268v1
- synphot: https://synphot.readthedocs.io/en/latest/
- pyMGal: https://arxiv.org/pdf/2507.00123v1
- RealSim-IFS: https://arxiv.org/pdf/2205.15335

and maybe to a lesser degree:
- simspin: https://ui.adsabs.harvard.edu/abs/2023PASA...40...48H/abstract
- popkinmocks: https://joss.theoj.org/papers/10.21105/joss.05225

This is a list I found, but there are likely more. Though I don't think any of them cover the same breadth, they still are important to put Synthesizer in context. For further context, it might be worth mentioning in the introduction that while the availability of public codes for synthetic observations has been recently growing, there is a long history of numerical simulators creating synthetic observations of their sims (I found this, but there's certainly more: https://arxiv.org/abs/1411.3717).

Other minor comments so far:
- It's a shame you don't mention SBI and multi-modal learning in the statement of need. Synthesizer seems perfectly designed for these applications (and I see you already have a paper on the subject).
- It might be good to mention the companion paper (https://arxiv.org/pdf/2508.03888v3) in the "Package overview" section to close the loop on the software paper, documentation, science paper combo
