### 2025-12-06 #9436 @TobiBu
Hi everyone,

I am just checking in on this. Thanks a lot @ConnorStoneAstro and @steven-murray for all your work and effort you put in so far.

@WillJRoper, there are already a couple comments to respond to. Then we can iterate to finalise the last few points on the reviewer checklists.
