### 2025-12-30 #9436 @steven-murray
Thanks @WillJRoper -- I made two small issues on the repo, but otherwise this looks good to me!
