### 2026-01-16 #9436 @TobiBu
Hi everyone,

that sounds awesome!
Thanks everyone for your efforts.

@ConnorStoneAstro there is indeed not much to do for you right now.

Looking forward to the final steps.
