### 2026-02-12 #9436 @TobiBu
Hi @WillJRoper,

what's the status of the paper update and the PR you mentioned [here](https://github.com/openjournals/joss-reviews/issues/9436#issuecomment-3761728665)
