### 2026-02-13 #9436 @WillJRoper
Sorry for the delay, all. The PR with the performance documentation has been merged (and there is actually a further expansion incoming beyond what was asked). As you can see in the above, I have also addressed @ConnorStoneAstro's comments on the paper itself and regenerated it.
