### 2026-02-13 #9436 @WillJRoper
Apologies, a collaborator just highlighted that some of the citations were out of order. Regenerating the fixed version above.
