### 2026-02-13 #9436 @ConnorStoneAstro
Hi @WillJRoper, I see you've updated the performance scaling tests and they look really good! However, I had asked in my original comment (https://github.com/openjournals/joss-reviews/issues/9436#issuecomment-3582237959) for you to compare with an external synthetic observation code. I checked the [Performance section](https://synthesizer-project.github.io/synthesizer/performance/performance.html) of the updated website and the [performance PR](https://github.com/synthesizer-project/synthesizer/pull/1086), but I didn't see such a comparison. For example, I imagine many potential users would be interested in how performance compares with the PROSPECTOR forward modelling code, which uses parametric representations of galaxies. The FORECAST code compares well with the particle modelling approach and would be a good test too. Is this still in progress? Or do you believe there is some obstacle to making a fair comparison?

The updated paper looks great! I really like the new "related packages" section. My one comment is that in the "Statement of Need" section you may want to explicitly state if Synthesizer fits in the "dust radiative transfer" or "bespoke pipelines" paradigm (or a third option).
