### 2026-02-13 #9436 @WillJRoper
>Hi @WillJRoper, I see you've updated the performance scaling tests and they look really good! However, I had asked in my original comment (https://github.com/openjournals/joss-reviews/issues/9436#issuecomment-3582237959) for you to compare with an external synthetic observation code. I checked the [Performance section](https://synthesizer-project.github.io/synthesizer/performance/performance.html) of the updated website and the https://github.com/synthesizer-project/synthesizer/pull/1086, but I didn't see such a comparison. For example, I imagine many potential users would be interested in how performance compares with the PROSPECTOR forward modelling code, which uses parametric representations of galaxies. The FORECAST code compares well with the particle modelling approach and would be a good test too. Is this still in progress? Or do you believe there is some obstacle to making a fair comparison?

Hi @ConnorStoneAstro, I apologise, I totally overlooked that part of your message. However, I'm somewhat disinclined to delve into comparisons on that level with other codes as it seems beyond the scope of the documentation. Furthermore, none of these softwares provide their own performance diagnostics in as much detail as the docs we have added (as far as I am aware ). The goal of the performance section was to show a user what to expect when they run the code in a particular situation. Even so, making a fair comparison between other codes might be a can of worms within itself to ensure we are actually comparing similar workflows and modelling choices and it's not a can of worms I am keen to open.

While I agree it would be interesting to explore, it could be its own comparison paper project in its own right, in which you could ensure like-for-like comparisons and compare performance and output fidelity.

>The updated paper looks great! I really like the new "related packages" section. My one comment is that in the "Statement of Need" section you may want to explicitly state if Synthesizer fits in the "dust radiative transfer" or "bespoke pipelines" paradigm (or a third option).

I'll get that changed ASAP.
