### 2026-02-13 #9436 @ConnorStoneAstro
Hi @WillJRoper, having done my own performance comparisons on other projects, I can say it is indeed a big can of worms. Given the modelling complexity of these synthetic observations, I agree it is fair to declare such a comparison out of scope.

I believe that crosses off all my comments! @TobiBu I am happy to recommend publication (once the minor change to the manuscript is complete).
