### 2026-02-15 #9436 @WillJRoper
Thanks @ConnorStoneAstro and @steven-murray. I have now addressed the final comments. All of us in the Synthesizer time massively appreciate your contributions! In the next comment, I will generate the final version of the paper for review.
