### 2026-02-17 #9436 @WillJRoper
Apologies, another proof with fixed affiliations and moving the "e.g." to the correct citation in a few places.
