### 2026-02-18 #9436 @TobiBu
Hi everyone,

this looks great.
I see that @steven-murray and @ConnorStoneAstro have completed their checklists and it looks like you agreed on a final version.

I will go over the most recent draft to check everything is in line with our [editorial guidelines](https://joss.readthedocs.io/en/latest/editing.html) and if so proceed with the publication process.
