### 2026-02-23 #9436 @TobiBu
Hi @WillJRoper,

I went one last time over the pdf and have one last comment:

1. In section Package Overview under components there is this sentence: `encapsulating physical properties, and emission and emission generation methods.` Is this intentional or is the double emission too much?
