### 2026-03-05 #9436 @TobiBu
> Hi [@WillJRoper](https://github.com/WillJRoper),
>
> there are a few more things left for the final publication. Can you please do the following so I can add the DOIs and versions to the final paper:
>
> * Double check authors and affiliations (including ORCIDs)
> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * Archive the release on Zenodo/figshare/etc and post the DOI here.

Hi @WillJRoper,

any updates on this? Did you create a new release and an archive version?
