### 2026-03-08 #9436 @WillJRoper
Hi @TobiBu

Sorry for the delay, same old job search excuse but I now have all of this done and dusted.

> - Double check authors and affiliations (including ORCIDs)

Done

> - Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Done -> [v1.0.0](https://github.com/synthesizer-project/synthesizer/releases/tag/v1.0.0)

> - Archive the release on Zenodo/figshare/etc and post the DOI here.

Done -> [10.5281/zenodo.18910099](https://doi.org/10.5281/zenodo.18910099)

>I went one last time over the pdf and have one last comment:
>
>In section Package Overview under components there is this sentence: encapsulating physical properties, and emission and emission generation methods. Is this intentional or is the double emission too much?

Indeed, this was a typo 😅 dealt with and regenerated above.
