### 2026-03-12 #9436 @TobiBu
Dear @WillJRoper,

thanks a lot for this. And sorry for my slow response I had to go back to all the other editors to understand better one line in my editor checklist:
`Check that the title, author list, version tag, and the license are correct in the archived release`

Looking at your [zenodo archive](https://doi.org/10.5281/zenodo.18910099) I saw that the title of the archive is slightly different to the paper and the author list has two names that are not on the paper.

I think this is because you used the GitHub zenodo integration tool.

Could you please adapt the metadata of the zenodo archive to adapt the title to the paper version? No need to create a new archive. In terms of the different author list, I guess it is ok if you briefly state here in this thread why the author list of the archive is different form that of the paper.
I think then we should be good to finalise this. Thanks a lot!

> Hi [@TobiBu](https://github.com/TobiBu)
>
> Sorry for the delay, same old job search excuse but I now have all of this done and dusted.
>
> > * Double check authors and affiliations (including ORCIDs)
>
> Done
>
> > * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
>
> Done -> [v1.0.0](https://github.com/synthesizer-project/synthesizer/releases/tag/v1.0.0)
>
> > * Archive the release on Zenodo/figshare/etc and post the DOI here.
>
> Done -> [10.5281/zenodo.18910099](https://doi.org/10.5281/zenodo.18910099)
>
> > I went one last time over the pdf and have one last comment:
> > In section Package Overview under components there is this sentence: encapsulating physical properties, and emission and emission generation methods. Is this intentional or is the double emission too much?
>
> Indeed, this was a typo 😅 dealt with and regenerated above.
