### 2026-03-12 #9436 @WillJRoper
Sorry @TobiBu, I did indeed automate it for future use. I should have thought that would be unhelpful 😅. We have had some new contributors since this process started, hence the different author list. Given that their work is in the archived release, I would actually prefer including them as authors wholesale (our policy within the project was that contribution == authorship on the paper). The fact they weren't added is a big oversight on my part in the first place!

Let me know if there is an issue with this. I will make the necessary ammendments at some point this morning.
