### 2026-03-12 #9436 @WillJRoper
Sorry, "this morning" proved ambitious. I have now edited the metadata of the Zenodo release to have a matching title and added the additional authors such that everything is consistent. I will regenerate the paper below.
