### 2026-03-12 #9436 @TobiBu
> Sorry, "this morning" proved ambitious. I have now edited the metadata of the Zenodo release to have a matching title and added the additional authors such that everything is consistent. I will regenerate the paper below.

Awesome. Now the only thing is Kartheik Iyer is missing on the zenodo metadata.
