### 2026-03-13 #9436 @WillJRoper
> Now the only thing is Kartheik Iyer is missing on the zenodo metadata.

Done
