### 2026-03-13 #9436 @TobiBu
Awesome! Thanks.
I guess then we are good.
Congrats to an awesome software project! Looking forward to more cool science with synthesizer.

> > Now the only thing is Kartheik Iyer is missing on the zenodo metadata.
>
> Done
