### 2026-03-13 #9436 @christopherlovell
@TobiBu apologies, there was one small error in the acknowledgements section that has now been corrected here https://github.com/synthesizer-project/synthesizer/blob/joss-paper/docs/joss_paper/paper.md
