### 2026-03-13 #9436 @TobiBu
@openjournals/aass-eics can you please make sure that the last changes in the acknowledgements made it into the final version before we finally publish.

> [@TobiBu](https://github.com/TobiBu) apologies, there was one small error in the acknowledgements section that has now been corrected here https://github.com/synthesizer-project/synthesizer/blob/joss-paper/docs/joss_paper/paper.md
