### 2026-03-17 #9436 @warrickball
Hi! There are a few checks I need to work through (and having steadily been doing so). This usually ends with my opening a PR to apply any final editorial touches, after which we'll do another test run and, if everything looks good, publish the paper.

I happen to be particularly busy at the moment but have worked through most of my usual checks. I'll try to wrap this up this evening, since I otherwise probably won't have another chance until the weekend.
