### 2026-03-17 #9436 @warrickball
Not a problem at all. I've just created the [relevant PR with various editorial tweaks](https://github.com/synthesizer-project/synthesizer/pull/1107). I'll be notified when that's merged and will then start another test run of the publication pipeline.
