### 2026-03-18 #9436 @warrickball
Thanks @WillJRoper, but now the affiliations aren't introduced in order again... The numbers should increase one by one as new affiliations appears in the author list, whereas we now have 1–7, 13, 9–12, and 8 (Jodrell Bank) is listed under the authors but AFAICT no author is actually affiliated with it. With apologies for my pedantry, could you fix those up, then I'll retest the publication pipeline? Thanks!
