### 2026-03-18 #9436 @WillJRoper
Don't apologise for the pedantry. a) serves me right for being lazy and rushing the addressing of the PR, and b) it made me realise the affiliations had gone seriously sideways. Not sure how or when that happened, possibly when someone updated their affiliations, I'm not sure. I have now taken the time to check each individually and ensure they are both correctly assigned and in the correct order. I am 100% sure these are now correct!
