### 2026-03-18 #9436 @warrickball
That's it @WillJRoper, your paper has been published in JOSS! 🎉

Thanks once again to @steven-murray and @ConnorStoneAstro for reviewing and @TobiBu for editing this submission! JOSS simply wouldn't be possible without its community of volunteers.
